# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Campwiz",
        "title": "Campwiz API",
        "contact": {
            "name": "Nokib Sarkar",
            "url": "https://github.com/nokibsarkar",
            "email": "nokibsarkar@gmail.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1"
    },
    "host": "localhost:8081",
    "basePath": "/api/v2",
    "paths": {
...
        "/round/": {
            "get": {
                "description": "get all rounds",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "List all rounds",
                "parameters": [
                    {
                        "type": "string",
                        "name": "campaignId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "next",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "prev",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "PENDING",
                            "IMPORTING",
                            "DISTRIBUTING",
                            "EVALUATING",
                            "REJECTED",
                            "CANCELLED",
                            "PAUSED",
                            "SCHEDULED",
                            "ACTIVE",
                            "COMPLETED"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "RoundStatusPending",
                            "RoundStatusImporting",
                            "RoundStatusDistributing",
                            "RoundStatusEvaluating",
                            "RoundStatusRejected",
                            "RoundStatusCancelled",
                            "RoundStatusPaused",
                            "RoundStatusScheduled",
                            "RoundStatusActive",
                            "RoundStatusCompleted"
                        ],
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseList-models_Round"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new round for a campaign",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Create a new round",
                "parameters": [
                    {
                        "description": "The round request",
                        "name": "roundRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.RoundRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-models_Round"
                        }
                    }
                }
            }
        },
        "/round/distribute/{roundId}": {
            "post": {
                "description": "Distribute evaluations to juries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Distribute evaluations to juries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The distribution request",
                        "name": "DistributionRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.DistributionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-models_Task"
                        }
                    }
                }
            }
        },
        "/round/import/{roundId}/commons": {
            "post": {
                "description": "The user would provide a round ID and a list of commons categories and the system would import images from those categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Import images from commons",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The import from commons request",
                        "name": "ImportFromCommons",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.ImportFromCommonsPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-models_Task"
                        }
                    }
                }
            }
        },
        "/round/import/{targetRoundId}/previous": {
            "post": {
                "description": "The user would provide a round ID and a list of scores and the system would import images from the previous round with those scores",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Import images from previous round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The target round ID, where the images will be imported",
                        "name": "targetRoundId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The import from previous round request",
                        "name": "ImportFromPreviousRoundPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.ImportFromPreviousRoundPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-models_Task"
                        }
                    }
                }
            }
        },
        "/round/{roundId}": {
            "get": {
                "description": "Get a round",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Get a round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-models_Round"
                        }
                    }
                }
            },
            "post": {
                "description": "Update the details of a round",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Update the details of a round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The round request",
                        "name": "roundRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.RoundRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-models_Round"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a round",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Delete a round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-routes_RoundDeletedResponse"
                        }
                    }
                }
            }
        },
        "/round/{roundId}/next/public": {
            "get": {
                "description": "Get the next public submission for a jury",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Get the next public submission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-models_Submission"
                        }
                    }
                }
            }
        },
        "/round/{roundId}/results/summary": {
            "get": {
                "description": "Get results of a round",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Get results of a round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseList-models_EvaluationResult"
                        }
                    }
                }
            }
        },
        "/round/{roundId}/results/{format}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get results of a round",
                "produces": [
                    "application/json",
                    "text/csv"
                ],
                "tags": [
                    "Round"
                ],
                "summary": "Get results of a round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "csv",
                            "json"
                        ],
                        "type": "string",
                        "description": "The format of the results",
                        "name": "format",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "next",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "prev",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "ARTICLE",
                                "BITMAP",
                                "AUDIO",
                                "VIDEO",
                                "PDF"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseList-models_SubmissionResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ResponseList-models_Round": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Round"
                    }
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseSingle-models_Round": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Round"
                }
            }
        },
...
        "models.ResponseSingle-models_Task": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Task"
                }
            }
        },
...
        "services.DistributionRequest": {
            "type": "object",
            "properties": {
                "juries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "services.ImportFromCommonsPayload": {
            "type": "object",
            "required": [
                "categories"
            ],
            "properties": {
                "categories": {
                    "description": "Categories from which images will be fetched",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "services.ImportFromPreviousRoundPayload": {
            "type": "object",
            "required": [
                "roundId",
                "scores"
            ],
            "properties": {
                "roundId": {
                    "description": "RoundID from which images will be fetched",
                    "type": "string"
                },
                "scores": {
                    "description": "Scores of the images to be fetched",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ScoreType"
                    }
                }
            }
        },
...
    }
}